@@ -1,12 +1,15 @@
 plugins {
-	id 'org.springframework.boot' version '2.7.3'
-	id 'io.spring.dependency-management' version '1.0.13.RELEASE'
 	id 'java'
+	id 'org.springframework.boot' version '3.1.3'
+	id 'io.spring.dependency-management' version '1.1.3'
 }
 
 group = 'oit.is.inudaisuki'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+
+java {
+	sourceCompatibility = '17'
+}
 
 repositories {
 	mavenCentral()
@@ -14,19 +17,30 @@ repositories {
 
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-security'
-  implementation 'org.springframework.boot:spring-boot-starter-jetty'
-	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+	implementation 'org.springframework.boot:spring-boot-starter-jetty'
+  implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter:2.2.2'
-	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5'
+	implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter:3.0.2'
+	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
 	runtimeOnly 'com.h2database:h2'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testImplementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter-test:3.0.2'
+  //testImplementation 'org.eclipse.jetty:jetty-server:11.0.14'
+  //testImplementation 'jakarta.servlet:jakarta.servlet-api:6.0.0'
 	testImplementation 'org.springframework.security:spring-security-test'
+}
+
+ext {
+    set("jakarta-servlet.version", '5.0.0')
 }
 
 tasks.named('test') {
 	useJUnitPlatform()
 }
+
+// tasks.withType(JavaCompile) {
+//     options.encoding = 'UTF-8'
+// }
 
 configurations {
   implementation.exclude group: 'org.springframework.boot', module: 'spring-boot-starter-tomcat'
