@@ -1,51 +1,17 @@
 package oit.is.inudaisuki.springboot_samples.security;
 
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.User.UserBuilder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 
+@Configuration
 @EnableWebSecurity
 public class Sample3AuthConfiguration {
-
-  /**
-   * 認証処理に関する設定（誰がどのようなロールでログインできるか）
-   *
-   * @return
-   */
-  @Bean
-  public InMemoryUserDetailsManager userDetailsService() {
-    // このクラスの下部にあるPasswordEncoderを利用して，ユーザのビルダ（ログインするユーザを作成するオブジェクト）を作成する
-    UserBuilder users = User.builder();
-
-    // UserBuilder usersにユーザ名，パスワード，ロールを指定してbuildする
-    // このときパスワードはBCryptでハッシュ化されている．
-    // ハッシュ化されたパスワードを得るには，この授業のbashターミナルで下記のように末尾にユーザ名とパスワードを指定すると良い(要VPN)
-    // $ sshrun htpasswd -nbBC 10 user1 p@ss
-    UserDetails user1 = users
-        .username("user1")
-        .password("$2y$10$ngxCDmuVK1TaGchiYQfJ1OAKkd64IH6skGsNw1sLabrTICOHPxC0e")
-        .roles("USER")
-        .build();
-    UserDetails user2 = users
-        .username("user2")
-        .password("$2y$10$ngxCDmuVK1TaGchiYQfJ1OAKkd64IH6skGsNw1sLabrTICOHPxC0e")
-        .roles("USER")
-        .build();
-    UserDetails admin = users
-        .username("admin")
-        .password("$2y$10$ngxCDmuVK1TaGchiYQfJ1OAKkd64IH6skGsNw1sLabrTICOHPxC0e")
-        .roles("ADMIN")
-        .build();
-    // 生成したユーザをImMemoryUserDetailsManagerに渡す（いくつでも良い）
-    return new InMemoryUserDetailsManager(user1, user2, admin);
-  }
 
   /**
    * 認可処理に関する設定（認証されたユーザがどこにアクセスできるか）
@@ -56,28 +22,41 @@ public class Sample3AuthConfiguration {
    */
   @Bean
   public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-    // Spring Securityのフォームを利用してログインを行う（自前でログインフォームを用意することも可能）
-    http.formLogin();
-
-    // http://localhost:8000/sample3 で始まるURLへのアクセスはログインが必要
-    // mvcMatchers().authenticated()がmvcMatchersに指定されたアクセス先に認証処理が必要であることを示す
-    // authenticated()の代わりにpermitAll()と書くと認証不要となる
-    http.authorizeHttpRequests()
-        .mvcMatchers("/sample3/**").authenticated();
-
-    http.logout().logoutSuccessUrl("/"); // ログアウト時は "http://localhost:8000/" に戻る
+    http.formLogin(login -> login
+        .permitAll())
+        .logout(logout -> logout
+            .logoutUrl("/logout")
+            .logoutSuccessUrl("/")) // ログアウト後に / にリダイレクト
+        .authorizeHttpRequests(authz -> authz
+            .requestMatchers("/*").permitAll() // すべてのユーザにアクセスを許可
+            .requestMatchers("/sample3/**").authenticated() //
+            .anyRequest().authenticated()); // 上記以外のページは認証が必要
     return http.build();
   }
 
   /**
+   * 認証処理に関する設定（誰がどのようなロールでログインできるか）
    *
-   * UserBuilder users = User.builder();で利用するPasswordEncoderを指定する．
-   *
-   * @return BCryptPasswordEncoderを返す
+   * @return
    */
   @Bean
-  PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder();
+  public InMemoryUserDetailsManager userDetailsService() {
+
+    // ユーザ名，パスワード，ロールを指定してbuildする
+    // このときパスワードはBCryptでハッシュ化されているため，{bcrypt}とつける
+    // ハッシュ化せずに平文でパスワードを指定する場合は{noop}をつける
+    // ハッシュ化されたパスワードを得るには，この授業のbashターミナルで下記のように末尾にユーザ名とパスワードを指定すると良い(要VPN)
+    // $ sshrun htpasswd -nbBC 10 user1 p@ss
+
+    UserDetails user1 = User.withUsername("user1")
+        .password("{bcrypt}$2y$10$ngxCDmuVK1TaGchiYQfJ1OAKkd64IH6skGsNw1sLabrTICOHPxC0e").roles("USER").build();
+    UserDetails user2 = User.withUsername("user2")
+        .password("{bcrypt}$2y$10$ngxCDmuVK1TaGchiYQfJ1OAKkd64IH6skGsNw1sLabrTICOHPxC0e").roles("USER").build();
+    UserDetails admin = User.withUsername("admin")
+        .password("{bcrypt}$2y$10$ngxCDmuVK1TaGchiYQfJ1OAKkd64IH6skGsNw1sLabrTICOHPxC0e").roles("ADMIN").build();
+
+    // 生成したユーザをImMemoryUserDetailsManagerに渡す（いくつでも良い）
+    return new InMemoryUserDetailsManager(user1, user2, admin);
   }
 
 }
